+++ test/unit/message_handler_spec.js
@@ -17,7 +17,7 @@ import {
   AbortException,
   UnknownErrorException,
 } from "../../src/shared/util.js";
-import { LoopbackPort } from "../../src/display/api.js";
+import { LoopbackPort } from "../../src/display/api_utils.js";
 import { MessageHandler } from "../../src/shared/message_handler.js";
 
 describe("message_handler", function () {
